server/pyproject.toml
[project]
name = "brainrotstudy"
version = "2.0.0"
description = "Turn PDFs, slides, or topics into goated 60-90s vertical study videos."
requires-python = ">=3.11"
readme = "../README.md"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.8",
    "pydantic-settings>=2.5",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "pymupdf>=1.24",
    "python-pptx>=1.0",
    "gtts>=2.5",
    "pillow>=10.3",
    "anthropic>=0.40",
    "google-genai>=0.3",
    "openai>=1.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.32",
    "ruff>=0.7",
]

[project.scripts]
brainrotstudy = "brainrotstudy.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brainrotstudy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"
